feat: refactor side panels and enabled overlay panels
