+++ src/sass/_side_panel.scss
@@ -1,0 +1,5 @@
+.is-overlay {
+  height: 100vh;
+  position: absolute;
+  z-index: 103 !important;
+}
