+++ src/sass/styles.scss
@@ -88,6 +88,7 @@ $border-thin: 1px solid $color-mid-light !default;
 @import "scrollable_table";
 @import "selectable_main_table";
 @import "settings_page";
+@import "side_panel";
 @import "snapshots";
 @import "status_bar";
 @import "storage_detail_overview";
